archive/scu_outreach_kit/requirements.txt
jinja2>=3.1.0
pandas>=2.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
pyyaml>=6.0
reportlab>=4.0.0
markdown>=3.5.0